Rename the auth QR plugin to avoid conflicts
