+++ mobile/apps/auth/lib/utils/gallery_import_util.dart
@@ -4,7 +4,7 @@ import 'package:backup_exclusion/backup_exclusion.dart';
 import 'package:ente_auth/l10n/l10n.dart';
 import 'package:ente_auth/models/code.dart';
 import 'package:ente_auth/utils/dialog_util.dart';
-import 'package:ente_qr/ente_qr.dart';
+import 'package:ente_auth_qr/ente_qr.dart';
 import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:logging/logging.dart';
