+++ mobile/apps/auth/plugins/qr/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.0.1
 
-* Initial release of ente_qr plugin
+* Initial release of ente_auth_qr plugin
 * Support for scanning QR codes from image files
 * Android implementation using ZXing library
 * iOS implementation using Core Image framework
