+++ mobile/apps/auth/plugins/qr/android/build.gradle
@@ -1,4 +1,4 @@
-group 'io.ente.auth.ente_qr'
+group 'io.ente.auth.ente_auth_qr'
 version '1.0-SNAPSHOT'
 
 buildscript {
@@ -25,7 +25,7 @@ apply plugin: 'com.android.library'
 apply plugin: 'kotlin-android'
 
 android {
-    namespace 'io.ente.auth.ente_qr'
+    namespace 'io.ente.auth.ente_auth_qr'
     compileSdk 36
 
     compileOptions {
