+++ mobile/apps/auth/plugins/qr/android/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'ente_qr'
+rootProject.name = 'ente_auth_qr'
