+++ mobile/apps/auth/plugins/qr/android/src/main/kotlin/io/ente/auth/ente_auth_qr/EnteQrPlugin.kt
@@ -1,4 +1,4 @@
-package io.ente.auth.ente_qr
+package io.ente.auth.ente_auth_qr
 
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
