+++ mobile/apps/auth/plugins/qr/lib/ente_qr.dart
@@ -1,4 +1,4 @@
-import 'package:ente_qr/ente_qr_platform_interface.dart';
+import 'package:ente_auth_qr/ente_qr_platform_interface.dart';
 
 export 'ente_qr_platform_interface.dart'
     show QrDetection, QrScanResult, QrScanResults;
