+++ mobile/apps/auth/plugins/qr/lib/ente_qr_platform_interface.dart
@@ -1,4 +1,4 @@
-import 'package:ente_qr/ente_qr_method_channel.dart';
+import 'package:ente_auth_qr/ente_qr_method_channel.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 /// The result of QR code scanning
