+++ mobile/apps/auth/plugins/qr/pubspec.yaml
@@ -1,4 +1,4 @@
-name: ente_qr
+name: ente_auth_qr
 description: "A QR code reader plugin for Ente."
 version: 0.0.1
 homepage: https://ente.com
@@ -23,7 +23,7 @@ flutter:
   plugin:
     platforms:
       android:
-        package: io.ente.auth.ente_qr
+        package: io.ente.auth.ente_auth_qr
         pluginClass: EnteQrPlugin
       ios:
         pluginClass: EnteQrPlugin
