+++ mobile/apps/auth/pubspec_overrides.yaml
@@ -1,4 +1,4 @@
-# melos_managed_dependency_overrides: ente_accounts,ente_base,ente_configuration,ente_events,ente_lock_screen,ente_logging,ente_network,ente_strings,ente_ui,ente_utils,scoped_dir_access,ente_crypto_api,ente_crypto_dart_adapter,ente_pure_utils,backup_exclusion,ente_qr
+# melos_managed_dependency_overrides: ente_accounts,ente_base,ente_configuration,ente_events,ente_lock_screen,ente_logging,ente_network,ente_strings,ente_ui,ente_utils,scoped_dir_access,ente_crypto_api,ente_crypto_dart_adapter,ente_pure_utils,backup_exclusion
 dependency_overrides:
   scoped_dir_access:
     path: ../../packages/scoped_dir_access
@@ -30,5 +30,3 @@ dependency_overrides:
     path: ../../packages/utils
   backup_exclusion:
     path: ../../packages/backup_exclusion
-  ente_qr:
-    path: ../photos/plugins/ente_qr
